Rename config to cli_parser. Still ugly, but as for me a bit better
