+++ .isort.cfg
@@ -10,4 +10,4 @@ ensure_newline_before_comments=True
 line_length=88
 split_on_trailing_comma=True
 known_third_party=fastapi,bittensor
-known_first_party=_validator,_miner,constants,utils,execution_layer,deployment_layer,miner,validator,protocol,models,constants,config
+known_first_party=_validator,_miner,constants,utils,execution_layer,deployment_layer,miner,validator,protocol,models,constants,cli_parser
