+++ neurons/_validator/validator_session.py
@@ -4,7 +4,7 @@ import sys
 
 import bittensor as bt
 
-import config
+import cli_parser
 from _validator.config import ValidatorConfig
 from _validator.core.validator_loop import ValidatorLoop
 from utils import clean_temp_files
@@ -12,7 +12,7 @@ from utils import clean_temp_files
 
 class ValidatorSession:
     def __init__(self):
-        self.config = ValidatorConfig(config.config)
+        self.config = ValidatorConfig(cli_parser.config)
         self.validator_loop = ValidatorLoop(self.config)
 
     def run(self):
